auth/requirements.txt
PyJWT>=2.8.0
